Module fwg::graphics::ui::widgets::prelude

Re-exports

pub use crate::gen_id;

pub use crate::graphics::ui::widgets::Widget;

Structs

Align

A Widget which wraps a contiained Widget to automatically align it within the available space.

Button

A Button is a UI widget which can fire a message when clicked.

Col

A Col is a collection of wigets which is arranged in a single horizontal row.

Container

A generic container for another Widget. Containers have margin, padding, and a border akin to the standard CSS box model.

Element

An Element is a type-erased widget. Elements allow UI objects to hold a variety of Widget implementations and dynamically dispatch function calls as needed.

Font

This struct contains all of the information required to render rasterized glyphs on screen.

HSplit

A widget which splits the available area in half horizontally for a left side and right side widget.

Id

This type represents a unique ID for User Interface elements.

Label

Row

A Row is a collection of wigets which is arranged in a single horizontal row.

Slider

Window

A Window is a collapsable panel with a title button which toggles the visibility of the contents.

Enums

Axis

Used to represent a layout axis for a list of dimensions.

Constraint

Constraints can be applied to the width and height of containers to limit their size.

HAlignment

Represents the Horizontal Alignment for a Widget wrapped by Align.

Justify

SpaceBetween

VAlignment

Represents the Vertical Alignment for a Widget wrapped by Align.

Traits

WithContainer

Define an extra associated method which automatically wraps any widget with a container.

Functions

align

Wrap a widget with an align widget.

button

Wrap the given widget into an interactive button.

col

Create a column of widgets.

hsplit

id_hash

Generate a simple hash from a given string at compile time. This is used by the gen_id!() macro to generate an ID number based on the file, line, and column number of the id.

label

Create a text label.

row

Create a row of widgets.

slider

text_button

Create a text-button.