Trait fwg::graphics::vulkan::VulkanDebug

pub trait VulkanDebug {
    fn set_debug_name(
        &self, 
        debug_name: impl Into<String>
    ) -> Result<(), VulkanDebugError>;
}

Types which implement this trait can set their name in the Vulkan validation layer logs.

Required methods

fn set_debug_name(
&self,
debug_name: impl Into<String>
) -> Result<(), VulkanDebugError>

Set the debug name for this resource in Vulkan validation layer logs.